pyproject.toml
[tool.poetry]
name = "learning-platform"
version = "0.1.0"
description = ""
authors = ["Sergei Gorlanov <itserg90@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
django = "^5.0.6"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
pillow = "^10.4.0"
black = "^24.4.2"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.0"
drf-yasg = "^1.21.7"
stripe = "^10.3.0"
forex-python = "^1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
